EOS/requirements.txt
# EOS — Python dependencies
# Install: pip install -r requirements.txt

# Core web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0

# Model & inference
# (llama-server binaries are bundled — no Python llama.cpp needed)

# Memory & retrieval
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Audio
sounddevice>=0.4.6
numpy>=1.26.0
faster-whisper>=1.0.0
# piper-tts: installed via Piper binary (bundled) — no pip package needed

# Vision / screen capture
mss>=9.0.1
Pillow>=10.3.0
opencv-python-headless>=4.9.0

# Discord
discord.py>=2.3.2

# Google Workspace
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.128.0

# Utilities
python-multipart>=0.0.9      # file upload support for FastAPI
websockets>=12.0             # WebSocket support
psutil>=5.9.0                # hardware sensors (CPU, RAM, disk)
PyNaCl>=1.5.0                # Discord voice encryption (required by discord.py)
davey>=0.1.0                 # Discord DAVE E2EE voice protocol (required by discord.py 2.4+)
